+++ lib/plug/accesslog/formatter/remote_ip_address.ex
@@ -1,15 +1,15 @@
-defmodule Plug.AccessLog.Formatter.RemoteHostname do
+defmodule Plug.AccessLog.Formatter.RemoteIPAddress do
   @moduledoc """
-  Determines remote hostname.
+  Determines remote ip address.
   """
 
   @doc """
   Appends to log output.
   """
   @spec append(String.t, Plug.Conn.t) :: String.t
-  def append(message, conn), do: message <> remote_hostname(conn)
+  def append(message, conn), do: message <> remote_ip_address(conn)
 
-  defp remote_hostname(conn) do
+  defp remote_ip_address(conn) do
     conn.remote_ip
     |> :inet_parse.ntoa()
     |> to_string()